+++ app/Providers/RouteServiceProvider.php
@@ -25,7 +25,8 @@ class RouteServiceProvider extends ServiceProvider
     public function boot(): void
     {
         RateLimiter::for(
-            'api', function (Request $request) {
+            'api',
+            function (Request $request) {
                 return Limit::perMinute(60)->by($request->user()?->id ?: $request->ip());
             }
         );
@@ -38,7 +39,7 @@ class RouteServiceProvider extends ServiceProvider
                     ->group(base_path('routes/api.php'));
 
                 Route::middleware('web')
-                ->group(base_path('routes/web.php'));
+                    ->group(base_path('routes/web.php'));
             }
         );
     }
