+++ routes/api/order.php
@@ -1,0 +1,6 @@
+<?php
+
+use App\Http\Controllers\Api\OrderController;
+use Illuminate\Support\Facades\Route;
+
+Route::post('/', [OrderController::class, 'store']);
